fix: deleted comments and renamed consts routes
