+++ src/app/constants/routes/index.ts
@@ -1,2 +1,2 @@
 export * from './auth'
-export * from './sidebar'
+export * from './menu-navigation'
